requirements.txt
pandas==1.3.5
python-dotenv==0.19.1